requirements.txt
pandas == 2.3.3
numpy == 2.26
xgboost == 3.1.1
opencv-python == 4.12.0.88
pillow == 11.1.0
joblib